Falta el ejercicio 6

+++ Logica/Incidencia.cs
@@ -15,7 +15,7 @@ namespace Logica
     //Punto 8
     public abstract class Incidencia
     {
-        public int MinutoDeJuego { get; set; }      
+        public short MinutoDeJuego { get; set; }      
         
         public abstract string ObtenerDescripcionIncidencia();
     }
